+++ CHANGELOG.md
@@ -72,6 +72,7 @@ and disable fields when "Settings (Inventory): Create, edit and delete HRID hand
 * Update required permissions for quick export of instances. Refs UIIN-3003.
 * Add `showSortIndicator` prop to `SearchAndSort` to display sort indicator next to the column names, and add correct `nonInteractiveHeaders`. Refs UIIN-2986.
 * Change casing of `Linked data editor` label. Refs UIIN-3007.
+* MARC bib/holdings | View Source: Add Edit and Export actions. Refs UIIN-3012.
 
 ## [11.0.4](https://github.com/folio-org/ui-inventory/tree/v11.0.4) (2024-04-30)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v11.0.3...v11.0.4)